+++ lib/rubocop/cop/layout/extra_spacing.rb
@@ -68,33 +68,41 @@ module RuboCop
         def check_tokens(ast, token1, token2)
           return if token2.type == :tNL
 
-          if force_equal_sign_alignment? &&
-             @asgn_tokens.include?(token2) &&
-             (@asgn_lines.include?(token2.line - 1) ||
-              @asgn_lines.include?(token2.line + 1))
+          if force_equal_sign_alignment? && @asgn_tokens.include?(token2)
             check_assignment(token2)
           else
             check_other(token1, token2, ast)
           end
         end
 
+        # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
         def check_assignment(token)
-          assignment_line = ''
-          message = ''
-          if should_aligned_with_preceding_line?(token)
-            assignment_line = processed_source.preceding_line(token)
-            message = format(MSG_UNALIGNED_ASGN, location: 'preceding')
-          else
-            assignment_line = processed_source.following_line(token)
-            message = format(MSG_UNALIGNED_ASGN, location: 'following')
-          end
+          token_line_indent              = processed_source
+                                           .line_indentation(token.line)
+
+          preceding_line_range           = token.line.downto(1)
+          preceding_assignment_lines     = \
+            relevant_assignment_lines(preceding_line_range)
+          preceding_relevant_line_number = preceding_assignment_lines[1]
+
+          return unless preceding_relevant_line_number
+
+          preceding_relevant_indent = \
+            processed_source
+            .line_indentation(preceding_relevant_line_number)
+
+          return if preceding_relevant_indent < token_line_indent
+
+          assignment_line = processed_source
+                            .lines[preceding_relevant_line_number - 1]
+          message         = format(MSG_UNALIGNED_ASGN, location: 'preceding')
+
+          return unless assignment_line
           return if aligned_assignment?(token.pos, assignment_line)
+
           add_offense(token.pos, location: token.pos, message: message)
         end
-
-        def should_aligned_with_preceding_line?(token)
-          @asgn_lines.include?(token.line - 1)
-        end
+        # rubocop:enable Metrics/AbcSize, Metrics/MethodLength
 
         def check_other(token1, token2, ast)
           extra_space_range(token1, token2) do |range|
@@ -173,7 +181,7 @@ module RuboCop
         end
 
         def align_equal_signs(range, corrector)
-          lines  = contiguous_assignment_lines(range)
+          lines  = all_relevant_assignment_lines(range.line)
           tokens = @asgn_tokens.select { |t| lines.include?(t.line) }
 
           columns  = tokens.map { |t| align_column(t) }
@@ -194,18 +202,47 @@ module RuboCop
           end
         end
 
-        def contiguous_assignment_lines(range)
-          result = [range.line]
+        def all_relevant_assignment_lines(line_number)
+          last_line_number = processed_source.lines.size
 
-          range.line.downto(1) do |lineno|
-            @asgn_lines.include?(lineno) ? result << lineno : break
-          end
-          range.line.upto(processed_source.lines.size) do |lineno|
-            @asgn_lines.include?(lineno) ? result << lineno : break
-          end
-
-          result.sort!
+          (
+            relevant_assignment_lines(line_number.downto(1)) +
+            relevant_assignment_lines(line_number.upto(last_line_number))
+          )
+            .uniq
+            .sort
         end
+
+        # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity
+        # rubocop:disable Metrics/PerceivedComplexity, Metrics/MethodLength
+        def relevant_assignment_lines(line_range)
+          result               = []
+          original_line_indent = processed_source
+                                 .line_indentation(line_range.first)
+          previous_line_indent_at_level = true
+
+          line_range.each do |line_number|
+            current_line_indent = processed_source.line_indentation(line_number)
+            blank_line          = processed_source.lines[line_number - 1].blank?
+
+            if (current_line_indent < original_line_indent && !blank_line) ||
+               (previous_line_indent_at_level && blank_line)
+              break
+            end
+
+            result << line_number if @asgn_lines.include?(line_number) &&
+                                     current_line_indent == original_line_indent
+
+            unless blank_line
+              previous_line_indent_at_level = \
+                current_line_indent == original_line_indent
+            end
+          end
+
+          result
+        end
+        # rubocop:enable Metrics/AbcSize, Metrics/CyclomaticComplexity
+        # rubocop:enable Metrics/PerceivedComplexity, Metrics/MethodLength
 
         def align_column(asgn_token)
           # if we removed unneeded spaces from the beginning of this =,