+++ spec/rubocop/cli/cli_autocorrect_spec.rb
@@ -31,6 +31,77 @@ RSpec.describe RuboCop::CLI, :isolated_environment do
     expect(IO.read('example.rb')).to eq(source)
   end
 
+  it 'plays nicely with default cops in complex ExtraSpacing scenarios' do
+    create_file('.rubocop.yml', <<-YAML.strip_indent)
+      # These cops change indentation and thus need disabling in order for the
+      # ExtraSpacing rules to apply to this scenario.
+
+      Layout/BlockAlignment:
+        Enabled: false
+
+      Layout/ExtraSpacingCop:
+        ForceEqualSignAlignment: true
+
+      Layout/MultilineMethodCallBraceLayout:
+        Enabled: false
+    YAML
+
+    source = <<-RUBY.strip_indent
+      def batch
+        @areas = params[:param].map do
+                     var_1 = 123_456
+                     variable_2 = 456_123
+                 end
+        @another = params[:param].map do
+                     char_1 = begin
+                                variable_1_1     = 'a'
+                                variable_1_20  = 'b'
+
+                                variable_1_300    = 'c'
+                                # A Comment
+                                variable_1_4000      = 'd'
+
+                                variable_1_50000     = 'e'
+                                puts 'a non-assignment statement without a blank line'
+                                some_other_length_variable     = 'f'
+                              end
+                     var_2 = 456_123
+                   end
+
+        render json: @areas
+      end
+    RUBY
+
+    create_file('example.rb', source)
+    expect(cli.run(['--auto-correct'])).to eq(1)
+
+    expect(IO.read('example.rb')).to eq(<<-RUBY.strip_indent)
+      def batch
+        @areas   = params[:param].map do
+                     var_1      = 123_456
+                     variable_2 = 456_123
+                   end
+        @another = params[:param].map do
+                     char_1 = begin
+                                variable_1_1  = 'a'
+                                variable_1_20 = 'b'
+
+                                variable_1_300  = 'c'
+                                # A Comment
+                                variable_1_4000 = 'd'
+
+                                variable_1_50000           = 'e'
+                                puts 'a non-assignment statement without a blank line'
+                                some_other_length_variable = 'f'
+                              end
+                     var_2  = 456_123
+                   end
+
+        render json: @areas
+      end
+    RUBY
+  end
+
   it 'does not correct SpaceAroundOperators in a hash that would be ' \
      'changed back' do
     create_file('.rubocop.yml', <<-YAML.strip_indent)